HZN-751: Allow Measurements API queries in template/generic info panel items
