+++ container/karaf/src/main/filtered-resources/etc/custom.properties
@@ -714,6 +714,8 @@ org.osgi.framework.system.packages.extra=org.apache.karaf.branding,\
         org.opennms.netmgt.model.alarm;version=${opennms.osgi.version},\
         org.opennms.netmgt.alarmd.api;version=${opennms.osgi.version},\
         org.opennms.netmgt.alarmd.api.support;version=${opennms.osgi.version},\
+        org.opennms.netmgt.measurements.api;version=${opennms.osgi.version},\
+        org.opennms.netmgt.measurements.model;version=${opennms.osgi.version},\
         org.opennms.netmgt.model.capsd;version=${opennms.osgi.version},\
         org.opennms.netmgt.model.discovery;version=${opennms.osgi.version},\
         org.opennms.netmgt.model.events;version=${opennms.osgi.version},\
