+++ features/measurements/impl/src/main/resources/META-INF/opennms/component-measurement.xml
@@ -2,13 +2,28 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
        xsi:schemaLocation="
            http://www.springframework.org/schema/beans
            http://www.springframework.org/schema/beans/spring-beans-4.0.xsd
            http://www.springframework.org/schema/context
-           http://www.springframework.org/schema/context/spring-context-4.0.xsd">
+           http://www.springframework.org/schema/context/spring-context-4.0.xsd
+           http://xmlns.opennms.org/xsd/spring/onms-osgi
+           http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd">
 
   <context:annotation-config />
   <context:component-scan base-package="org.opennms.netmgt.measurements"/>
+
+  <bean id="measurementFetchStrategyFactory" class="org.opennms.netmgt.measurements.api.MeasurementFetchStrategyFactory"/>
+  <bean id="filterEngine" class="org.opennms.netmgt.measurements.api.FilterEngine" />
+  <bean id="expressionEngine" class="org.opennms.netmgt.measurements.impl.JEXLExpressionEngine" />
+
+  <bean id="measurementsService" class="org.opennms.netmgt.measurements.api.DefaultMeasurementsService">
+    <constructor-arg ref="measurementFetchStrategy"/>
+    <constructor-arg ref="expressionEngine"/>
+    <constructor-arg ref="filterEngine"/>
+  </bean>
+
+  <onmsgi:service interface="org.opennms.netmgt.measurements.api.MeasurementsService" ref="measurementsService"/>
 
 </beans>
