+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/pom.xml
@@ -115,6 +115,12 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
+      <groupId>org.opennms.features.measurements</groupId>
+      <artifactId>org.opennms.features.measurements.api</artifactId>
+      <version>${project.version}</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
 	   <groupId>org.osgi</groupId>
 	   <artifactId>org.osgi.compendium</artifactId>
 	   <scope>provided</scope>
