+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/src/main/java/org/opennms/features/topology/plugins/topo/graphml/info/GenericInfoPanelItemProvider.java
@@ -53,6 +53,7 @@ import org.opennms.features.topology.api.topo.EdgeRef;
 import org.opennms.features.topology.api.topo.VertexRef;
 import org.opennms.features.topology.plugins.topo.graphml.GraphMLVertex;
 import org.opennms.netmgt.dao.api.NodeDao;
+import org.opennms.netmgt.measurements.api.MeasurementsService;
 import org.opennms.netmgt.model.OnmsNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -90,9 +91,12 @@ public class GenericInfoPanelItemProvider implements InfoPanelItemProvider {
 
     private final Jinjava jinjava;
 
-    public GenericInfoPanelItemProvider(NodeDao nodeDao) throws InstantiationException, IllegalAccessException {
+    private final MeasurementsService measurementsService;
+
+    public GenericInfoPanelItemProvider(NodeDao nodeDao, MeasurementsService measurementsService) throws InstantiationException, IllegalAccessException {
         this.jinjava = withClassLoaderFix(Jinjava::new);
         this.nodeDao = Objects.requireNonNull(nodeDao);
+        this.measurementsService = Objects.requireNonNull(measurementsService);
     }
 
     private class TemplateItem implements InfoPanelItem {
@@ -235,8 +239,7 @@ public class GenericInfoPanelItemProvider implements InfoPanelItemProvider {
                 .map(this::createVertexContext)
                 .ifPresent(context::putAll);
 
-        // TODO Measurements-API
-//        context.put("measurements", new MeasurementsWrapper(measurementsService));
+        context.put("measurements", new MeasurementsWrapper(measurementsService));
 
         return context;
     }
